printf: introduce a callback-driven printf function
This can be used to format strings without requiring a buffer large
enough to hold the entire result string. It's useful to implement
functions similar to printf() (i.e. to stdout) or fprintf(). When the
buffer gets full, the current data can be flushed to free the buffer,
and formatting continues.

This API/implementation is basically designed to append to a user
provided buffer. The user has full control over the buffer when calling
lin_bprintf(), and again when the callback is entered.

This is vaguely similar to how libcs typically implement both snprintf
and fprintf using the same printf core. Some of these expose this
functionality to API users by making FILE customizable, so fprintf()
would append to a user-provided buffer, plus full control over overflow
handling. But we don't want to provide FILE emulation, or have any heavy
weight "byte stream" abstractions around. We want a light weight API
without too many dependencies. (FILE can surely be implemented with very
low footprint, but it still introduces an entire class of APIs. Using
the libc's FILE is not an option, because the relevant FILE APIs are not
standard. Good job again, C committee.)

Passing pointer-to-pointers as buffer argument is a bit awkward, but in
the end I found it less awkward than some other approaches, such as
using a different struct with buffer pointers. I didn't want to have a
weird single-use struct floating around that would look like a
half-assed FILE. The current API is still very reasonable, just a bit
awkward.

It seems this got optimized towards reducing the number of callbacks.
This does not even invoke any callbacks if the buffer space is enough
(and thus is as efficient as lin_snprintf()). I'm not sure whether that
was a good focus, or how much the penalty incurred on various platforms
(from desktop to embedded) matters relative to typical snprintf uses.
But surely we don't want to invoke a callback for every character, and
reducing callbacks just seems like a good idea in general.

Other APIs would have been possible.

The FILE approach was discussed above. It is potentially the most
flexible one.

One idea was an out(char *buf, size_t size) callback. You could have
output %s arguments directly as separate out() call, without going
through an immediate buffer. The lin_bprintf() signature could have been
very simple. In the simplest case, you would pass no buffer, and a small
internal stack buffer would have been used to batch output of multiple
single characters. A slightly more complex improvement would require the
user to provide this scratch buffer, to reduce stack usage and to force
the user to make efficiency/memory usage tradeoffs. I discarded this
idea because it would have resulted in a higher number of callback
invocations depending on the format string. Or we would have needed
cleverer buffering to avoid out() calls with small sizes and merge them
in the scratch buffer. All in all this didn't look too promising, and
the user would possibly have been required to do their own additional
buffering. (I was thinking of fprintf()-like uses, where you don't want
to flush to disk that often).

Another idea was making the user pass in a temporary buffer (like with
snprintf), plus a flush callback if the buffer was overran. But I got
hung up on some details. For example, I was about to add a buffer offset
parameter, so the user could append to an existing buffer. Or the
callback could have had a way to modify the buffer pointers. It seemed
like a simple idea at first, but making it not suck at some important
use-cases was hard. Strictly speaking, the current API is a variation of
this, but fulfills all the requirements.