add rest of musl files / undo unneeded changes to it
This is now all of the source code/install of musl (as described in the
README), though still only the "needed" top-level dirs.

I'd really like this to be a sub-repo, but for now the syscall parts for
"redirection" need to be different. Since musl provides absolutely
nothing to do this without intrusive changes to the musl source code,
they're still in the risc-v arch include files.

Other unnecessary changes (some of them pretty weird) are just undone.